chore: Better File Structuring
