+++ luna@thanderoy.github.io/stylesheet.css
@@ -42,7 +42,7 @@
 /* Separator */
 .separator {
     height: 1px;
-    background-color: rgba(255 255 255 / 50%);
+    background-color: rgba(255, 255, 255, 0.5);
     margin: 8px 0;
 }
 
@@ -51,10 +51,11 @@
     border-radius: 4px;
     font-weight: bold;
     padding: 8px 16px;
-    background-color: rgba(255 255 255 / 10%);
+    background-color: rgba(255, 255, 255, 0.1);
     color: #fff;
+    transition-duration: 200ms;
 }
 
 .refresh-button:hover {
-    background-color: rgba(255 255 255 / 20%);
+    background-color: rgba(255, 255, 255, 0.2);
 }
